groupGeneric {pROC}	R Documentation

pROC Group Generic Functions

Description

Redefine base groupGeneric functions to handle auc and ci objects properly on operations and mathematical operations. Attributes are dropped so that the AUC/CI behaves as a numeric value/matrix, respectively. In the case of AUC, all attributes are dropped, while in CI only the CI-specific attributes are, keeping those necessary for the matrices.

Usage


Math(x, ...)
Ops(e1, e2)

Arguments

x, e1, e2
auc objects, or mixed numerics and auc objects.

...
further arguments passed to other Math methods.

See Also

groupGeneric, auc

Examples


data(aSAH)

# Create a roc object:
aucobj1 <- auc(roc(aSAH$outcome, aSAH$s100b))
aucobj2 <- auc(roc(aSAH$outcome, aSAH$wfns))

# Math
sqrt(aucobj1)
round(aucobj2, digits=1)

# Ops
aucobj1 * 2
2 * aucobj2
aucobj1 + aucobj2

# With CI
ciaucobj <- ci(aucobj1)
ciaucobj * 2
sqrt(ciaucobj)


[Package pROC version 1.18.0 Index]