print {pROC}	R Documentation

Print a ROC curve object

Description

This function prints a ROC curve, AUC or CI object and return it invisibly.

Usage


## S3 method for class 'roc'
print(x, digits=max(3, getOption("digits") - 3), call=TRUE, ...)
## S3 method for class 'multiclass.roc'
print(x, digits=max(3, getOption("digits") -
  3), call=TRUE, ...) 
## S3 method for class 'mv.multiclass.roc'
print(x, digits=max(3, getOption("digits") -
  3), call=TRUE, ...) 
## S3 method for class 'smooth.roc'
print(x, digits=max(3, getOption("digits") - 3),
call=TRUE, ...)
## S3 method for class 'auc'
print(x, digits=max(3, getOption("digits") - 3), ...)
## S3 method for class 'multiclass.auc'
print(x, digits=max(3, getOption("digits") - 3), ...)
## S3 method for class 'ci.auc'
print(x, digits=max(3, getOption("digits") - 3), ...)
## S3 method for class 'ci.thresholds'
print(x, digits=max(3, getOption("digits") - 3), ...)
## S3 method for class 'ci.se'
print(x, digits=max(3, getOption("digits") - 3), ...)
## S3 method for class 'ci.sp'
print(x, digits=max(3, getOption("digits") - 3), ...)
## S3 method for class 'ci.coords'
print(x, digits=max(3, getOption("digits") - 3), ...)

Arguments

x
a roc, auc or ci object, from the roc, auc or ci functions respectively.

call
if the call is printed.

digits
the number of significant figures to print. See signif for more details.

...
further arguments passed to or from other methods. In particular, print.roc calls print.auc and the print.ci variants internally, and a digits argument is propagated. Not used in print.auc and print.ci variants.

Value

These functions return the object they were passed invisibly.

See Also

roc, auc, ci, coords

Examples


data(aSAH)

# Print a roc object:
rocobj <- roc(aSAH$outcome, aSAH$s100b)
print(rocobj)

# Print a smoothed roc object
print(smooth(rocobj))

# implicit printing
 roc(aSAH$outcome, aSAH$s100b)

# Print an auc and a ci object, from the ROC object or calling
# the dedicated function:
print(rocobj$auc)
print(ci(rocobj))

[Package pROC version 1.18.0 Index]