smooth {pROC}	R Documentation

Smooth a ROC curve

Description

This function smoothes a ROC curve of numeric predictor. By default, a binormal smoothing is performed, but density or custom smoothings are supported.

Usage


smooth(...)
## Default S3 method:
smooth(...)
## S3 method for class 'roc'
smooth(roc,
method=c("binormal", "density", "fitdistr", "logcondens",
"logcondens.smooth"), n=512, bw = "nrd0", density=NULL,
density.controls=density, density.cases=density,
start=NULL, start.controls=start, start.cases=start, 
reuse.auc=TRUE, reuse.ci=FALSE, ...)
## S3 method for class 'smooth.roc'
smooth(smooth.roc, ...)

Arguments

roc, smooth.roc
a “roc” object from the roc function, or a “smooth.roc” object from the smooth function.

method
“binormal”, “density”, “fitdistr”, “logcondens”, “"logcondens.smooth"”.

n
the number of equally spaced points where the smoothed curve will be calculated.

bw
if method="density" and density.controls and density.cases are not provided, bw is passed to density to determine the bandwidth of the density Can be a character string (“nrd0”, “nrd”, “ucv”, “bcv” or “SJ”, but any name matching a function prefixed with “bw.” is supported) or a numeric value, as described in density. Defaults to “nrd0”.

density, density.controls, density.cases
if method="density", a numeric value of density (over the y axis) or a function returning a density (such as density. If method="fitdistr", a densfun argument for fitdistr. If the value is different for control and case observations, density.controls and density.cases can be employed instead, otherwise density will be propagated to both density.controls and density.cases.

start, start.controls, start.cases
if method="fitdistr", optionnal start arguments for . start.controls and start.cases allows to specify different distributions for controls and cases.

reuse.auc, reuse.ci
if TRUE (default for reuse.auc) and the “roc” objects contain “auc” or “ci” fields, re-use these specifications to regenerate auc or ci on the smoothed ROC curve with the original parameters. If FALSE, the object returned will not contain “auc” or “ci” fields. It is currently not possible to redefine auc and ci options directly: you need to call auc or ci later for that.

...
further arguments passed to or from other methods, and especially to density (only cut, adjust, and kernel, plus window for compatibility with S+) and fitdistr.

Details

If method="binormal", a linear model is fitted to the quantiles of the sensitivities and specificities. Smoothed sensitivities and specificities are then generated from this model on n points. This simple approach was found to work well for most ROC curves, but it may produce hooked smooths in some situations (see in Hanley (1988)).

With method="density", the density function is employed to generate a smooth kernel density of the control and case observations as described by Zhou et al. (1997), unless density.controls or density.cases are provided directly. bw can be given to specify a bandwidth to use with density. It can be a numeric value or a character string (“nrd0”, “nrd”, “ucv”, “bcv” or “SJ”, but any name matching a function prefixed with “bw.” is supported). In the case of a character string, the whole predictor data is employed to determine the numeric value to use on both controls and cases. Depending on your data, it might be a good idea to specify the kernel argument for density. By default, “gaussian” is used, but “epanechnikov”, “rectangular”, “triangular”, “biweight”, “cosine” and “optcosine” are supported. As all the kernels are symetrical, it might help to normalize the data first (that is, before calling roc), for example with quantile normalization:


    norm.x <- qnorm(rank(x)/(length(x)+1))
    smooth(roc(response, norm.x, ...), ...)
  

Additionally, density can be a function which must return either a numeric vector of densities over the y axis or a list with a “y” item like the density function. It must accept the following input:


    density.fun(x, n, from, to, bw, kernel, ...)
  

It is important to honour n, from and to in order to have the densities evaluated on the same points for controls and cases. Failing to do so and returning densities of different length will produce an error. It is also a good idea to use a constant smoothing parameter (such as bw) especially when controls and cases have a different number of observations, to avoid producing smoother or rougher densities.

If method="fitdistr", the fitdistr function from the MASS package is employed to fit parameters for the density function density with optionnal start parameters start. The density function are fitted separately in control (density.controls, start.controls) and case observations (density.cases, start.cases). density can be one of the character values allowed by fitdistr or a density function (such as dnorm, dweibull, ...).

The method="logcondens" and method="logcondens.smooth" use the logcondens package to generate a non smoothed or smoothed (respectively) log-concave density estimate of of the control and case observation with the logConROC function.

smooth.default forces the usage of the smooth function in the stats package, so that other code relying on smooth should continue to function normally.

Smoothed ROC curves can be passed to smooth again. In this case, the smoothing is not re-applied on the smoothed ROC curve but the original “roc” object will be re-used.

Note that a smooth.roc curve has no threshold.

Value

A list of class “smooth.roc” with the following fields:

sensitivities
the smoothed sensitivities defining the ROC curve.

specificities
the smoothed specificities defining the ROC curve.

percent
if the sensitivities, specificities and AUC are reported in percent, as defined in argument.

direction
the direction of the comparison, as defined in argument.

call
how the function was called. See match.call for more details.

smoothing.args
a list of the arguments used for the smoothing. Will serve to apply the smoothing again in further bootstrap operations.

auc
if the original ROC curve contained an AUC, it is computed again on the smoothed ROC.

ci
if the original ROC curve contained a CI, it is computed again on the smoothed ROC.

fit.controls, fit.cases
with method="fitdistr" only: the result of MASS's fitdistr function for controls and cases, with an additional “densfun” item indicating the density function, if possible as character.

logcondens
with method="logcondens" and method="logcondens.smooth" only: the result of logcondens's logConROC function.

model
with method="binormal" only: the linear model from lm used to smooth the ROC curve.

Attributes

Additionally, the original roc object is stored as a “roc” attribute.

Errors

The message “The 'density' function must return a numeric vector or a list with a 'y' item.” will be displayed if the density function did not return a valid output. The message “Length of 'density.controls' and 'density.cases' differ.” will be displayed if the returned value differ in length.

Binormal smoothing cannot smooth ROC curve defined by only one point. Any such attempt will fail with the error “ROC curve not smoothable (not enough points).”.

If the smooth ROC curve was generated by roc with density.controls and density.cases numeric arguments, it cannot be smoothed and the error “Cannot smooth a ROC curve generated directly with numeric 'density.controls' and 'density.cases'.” is produced.

fitdistr and density smoothing methods require a numeric predictor. If the ROC curve to smooth was generated with an ordered factor only binormal smoothing can be applied and the message “ROC curves of ordered predictors can be smoothed only with binormal smoothing.” is displayed otherwise.

fitdistr, logcondens and logcondens.smooth methods require additional packages. If not available, the following message will be displayed with the required command to install the package: “Package ? not available, required with method='?'. Please install it with 'install.packages("?")'. ”

References

James E. Hanley (1988) “The robustness of the “binormal” assumptions used in fitting ROC curves”. Medical Decision Making 8, 197–203.

Lutz Duembgen, Kaspar Rufibach (2011) “logcondens: Computations Related to Univariate Log-Concave Density Estimation”. Journal of Statistical Software, 39, 1–28. URL: jstatsoft.org/v39/i06.

Xavier Robin, Natacha Turck, Alexandre Hainard, et al. (2011) “pROC: an open-source package for R and S+ to analyze and compare ROC curves”. BMC Bioinformatics, 7, 77. DOI: doi: 10.1186/1471-2105-12-77.

Kaspar Rufibach (2011) “A Smooth ROC Curve Estimator Based on Log-Concave Density Estimates”. The International Journal of Biostatistics, 8, accepted. DOI: doi: 10.1515/1557-4679.1378. arXiv: 1103.1787.

William N. Venables, Brian D. Ripley (2002). “Modern Applied Statistics with S”. New York, Springer. Google books.

Kelly H. Zou, W. J. Hall and David E. Shapiro (1997) “Smooth non-parametric receiver operating characteristic (ROC) curves for continuous diagnostic tests”. Statistics in Medicine 18, 2143–2156. DOI: doi: 10.1002/(SICI)1097-0258(19971015)16:19<2143::AID-SIM655>3.0.CO;2-3.

See Also

roc

CRAN packages MASS and logcondens employed in this function.

Examples


data(aSAH)

##  Basic example
rocobj <- roc(aSAH$outcome, aSAH$s100b)
smooth(rocobj)
# or directly with roc()
roc(aSAH$outcome, aSAH$s100b, smooth=TRUE)

# plotting
plot(rocobj)
rs <- smooth(rocobj, method="binormal")
plot(rs, add=TRUE, col="green")
rs2 <- smooth(rocobj, method="density")
plot(rs2, add=TRUE, col="blue")
rs3 <- smooth(rocobj, method="fitdistr", density="lognormal")
plot(rs3, add=TRUE, col="magenta")
if (requireNamespace("logcondens")) {
rs4 <- smooth(rocobj, method="logcondens")
plot(rs4, add=TRUE, col="brown")
rs5 <- smooth(rocobj, method="logcondens.smooth")
plot(rs5, add=TRUE, col="orange")
}
legend("bottomright", legend=c("Empirical", "Binormal", "Density", "Log-normal",
                               "Log-concave density", "Smoothed log-concave density"),
       col=c("black", "green", "blue", "magenta", "brown", "orange"), lwd=2)

## Advanced smoothing

# if we know the distributions are normal with sd=0.1 and an unknown mean:
smooth(rocobj, method="fitdistr", density=dnorm, start=list(mean=1), sd=.1)
# different distibutions for controls and cases:
smooth(rocobj, method="fitdistr", density.controls="normal", density.cases="lognormal")

# with densities
bw <- bw.nrd0(rocobj$predictor)
density.controls <- density(rocobj$controls, from=min(rocobj$predictor) - 3 * bw,
                            to=max(rocobj$predictor) + 3*bw, bw=bw, kernel="gaussian")
density.cases <- density(rocobj$cases, from=min(rocobj$predictor) - 3 * bw,
                            to=max(rocobj$predictor) + 3*bw, bw=bw, kernel="gaussian")
smooth(rocobj, method="density", density.controls=density.controls$y, 
       density.cases=density.cases$y)
# which is roughly what is done by a simple:
smooth(rocobj, method="density")

## Not run: 
## Smoothing artificial ROC curves

rand.unif <- runif(1000, -1, 1)
rand.exp <- rexp(1000)
rand.norm <- 
rnorm(1000)

# two normals
roc.norm <- roc(controls=rnorm(1000), cases=rnorm(1000)+1, plot=TRUE)
plot(smooth(roc.norm), col="green", lwd=1, add=TRUE)
plot(smooth(roc.norm, method="density"), col="red", lwd=1, add=TRUE)
plot(smooth(roc.norm, method="fitdistr"), col="blue", lwd=1, add=TRUE)
if (requireNamespace("logcondens")) {
plot(smooth(roc.norm, method="logcondens"), col="brown", lwd=1, add=TRUE)
plot(smooth(roc.norm, method="logcondens.smooth"), col="orange", lwd=1, add=TRUE)
}
legend("bottomright", legend=c("empirical", "binormal", "density", "fitdistr",
                               "logcondens", "logcondens.smooth"), 
       col=c(par("fg"), "green", "red", "blue", "brown", "orange"), lwd=c(2, 1, 1, 1))
       
# deviation from the normality
roc.norm.exp <- roc(controls=rnorm(1000), cases=rexp(1000), plot=TRUE)
plot(smooth(roc.norm.exp), col="green", lwd=1, add=TRUE)
plot(smooth(roc.norm.exp, method="density"), col="red", lwd=1, add=TRUE)
# Wrong fitdistr: normality assumed by default
plot(smooth(roc.norm.exp, method="fitdistr"), col="blue", lwd=1, add=TRUE)
# Correct fitdistr
plot(smooth(roc.norm.exp, method="fitdistr", density.controls="normal",
            density.cases="exponential"), col="purple", lwd=1, add=TRUE)
if (requireNamespace("logcondens")) {
plot(smooth(roc.norm.exp, method="logcondens"), col="brown", lwd=1, add=TRUE)
plot(smooth(roc.norm.exp, method="logcondens.smooth"), col="orange", lwd=1, add=TRUE)
}
legend("bottomright", legend=c("empirical", "binormal", "density",
                               "wrong fitdistr", "correct fitdistr",
                               "logcondens", "logcondens.smooth"),
       col=c(par("fg"), "green", "red", "blue", "purple", "brown", "orange"), lwd=c(2, 1, 1, 1, 1))

# large deviation from the normality
roc.unif.exp <- roc(controls=runif(1000, 2, 3), cases=rexp(1000)+2, plot=TRUE)
plot(smooth(roc.unif.exp), col="green", lwd=1, add=TRUE)
plot(smooth(roc.unif.exp, method="density"), col="red", lwd=1, add=TRUE)
plot(smooth(roc.unif.exp, method="density", bw="ucv"), col="magenta", lwd=1, add=TRUE)
# Wrong fitdistr: normality assumed by default (uniform distributions not handled)
plot(smooth(roc.unif.exp, method="fitdistr"), col="blue", lwd=1, add=TRUE)
if (requireNamespace("logcondens")) {
plot(smooth(roc.unif.exp, method="logcondens"), col="brown", lwd=1, add=TRUE)
plot(smooth(roc.unif.exp, method="logcondens.smooth"), col="orange", lwd=1, add=TRUE)
}
legend("bottomright", legend=c("empirical", "binormal", "density",
                               "density ucv", "wrong fitdistr",
                               "logcondens", "logcondens.smooth"),
       col=c(par("fg"), "green", "red", "magenta", "blue", "brown", "orange"), lwd=c(2, 1, 1, 1, 1))

## End(Not run)

# 2 uniform distributions with a custom density function
unif.density <- function(x, n, from, to, bw, kernel, ...) {
  smooth.x <- seq(from=from, to=to, length.out=n)
  smooth.y <- dunif(smooth.x, min=min(x), max=max(x))
  return(smooth.y)
}
roc.unif <- roc(controls=runif(1000, -1, 1), cases=runif(1000, 0, 2), plot=TRUE)
s <- smooth(roc.unif, method="density", density=unif.density)
plot(roc.unif)
plot(s, add=TRUE, col="grey")

## Not run: 
# you can bootstrap a ROC curve smoothed with a density function:
ci(s, boot.n=100)

## End(Not run)

[Package pROC version 1.18.0 Index]